Resolved a small issue with recording

+++ src/main/java/com/sugarcrm/candybean/runner/TestRecorder.java
@@ -87,13 +87,13 @@ public class TestRecorder extends RunListener {
 	/**
 	 * The default file location to store test results in XML
 	 */
-	private static final String FAILED_TEST_RESULTS_XML = "./target/candybean-reports/results/failedTestResults.xml";
+	private static final String FAILED_TEST_RESULTS_XML = "./log/failedTestResults.xml";
 
 	/**
 	 * The default file location to create the html report of the failed test
 	 * results
 	 */
-	private static final String FAILED_RECORDING_REPORT_HTML = "./target/candybean-reports/reports/FailedRecordingReport.html";
+	private static final String FAILED_RECORDING_REPORT_HTML = "./log/FailedRecordingReport.html";
 
 	/**
 	 * The default file location of the surefire results directory
@@ -103,7 +103,7 @@ public class TestRecorder extends RunListener {
 	/**
 	 * The default path of where the candybean test results report will be generated
 	 */
-	private static final String CANDYBEAN_REPORT_PATH = "./target/candybean-reports/reports/CandybeanTestResults.html";
+	private static final String CANDYBEAN_REPORT_PATH = "./log/candybeanTestResults.html";
 
 	private static final String TEST_TEMPLATE_PATH = "./resources/html/testTemplate.html";
 
@@ -124,7 +124,7 @@ public class TestRecorder extends RunListener {
 		}
 		config = new Configuration(new File(
 				Utils.adjustPath(candybeanConfigStr)));
-		xmlFile = createFile(config.getValue("testResultsXMLPath", FAILED_TEST_RESULTS_XML));
+		xmlFile = createFile(config.getValue("testResultsXMLPath", FAILED_TEST_RESULTS_XML), false);
 	}
 
 	@Override
@@ -321,7 +321,7 @@ public class TestRecorder extends RunListener {
 			}
 			baseTemplate = baseTemplate.replace("${packageMarkup}", packageMarkup.toString());
 			FileWriter fstream = new FileWriter(createFile(config.getValue(
-					"testResultsReportPath", CANDYBEAN_REPORT_PATH)));
+					"testResultsReportPath", CANDYBEAN_REPORT_PATH), true));
 			BufferedWriter out = new BufferedWriter(fstream);
 			out.write(baseTemplate);
 			out.close();
@@ -361,13 +361,10 @@ public class TestRecorder extends RunListener {
 			}
 			reportTemplate = reportTemplate.replace("${recording.rows}", entryMarkup.toString());
 			String pathToRecordingReport = config.getValue("testResultsHtmlPath", FAILED_RECORDING_REPORT_HTML);
-			FileWriter fstream = new FileWriter(createFile(pathToRecordingReport));
+			FileWriter fstream = new FileWriter(createFile(pathToRecordingReport, true));
 			BufferedWriter out = new BufferedWriter(fstream);
 			out.write(reportTemplate);
 			out.close();
-		}else{
-			xmlFile.delete();
-			new File(config.getValue("testResultsHtmlPath", FAILED_RECORDING_REPORT_HTML)).delete();
 		}
 	}
 	
@@ -382,6 +379,7 @@ public class TestRecorder extends RunListener {
 		byte[] encoded = Files.readAllBytes(Paths.get(path));
 		return encoding.decode(ByteBuffer.wrap(encoded)).toString();
 	}
+
 	
 	/**
 	 * Creates a file at the given path
@@ -389,13 +387,19 @@ public class TestRecorder extends RunListener {
 	 * @return
 	 * @throws IOException
 	 */
-	private File createFile(String path) throws IOException{
+	private File createFile(String path, boolean replace) throws IOException{
 		File f = new File(path);
 		if (!f.getParentFile().exists()) {
 			f.getParentFile().mkdirs();
 		}
-		f.delete();
-		f.createNewFile();
+		if (f.exists()) {
+			if (replace) {
+				f.delete();
+				f.createNewFile();
+			}
+		} else {
+			f.createNewFile();
+		}
 		return f;
 	}
